Refactor logging
The directory structure of produced artifacts has been changed:
the root directory is configurable and defaults to the project
directory. Inside the root directory will be created the subdirectories
"results" and "runs", which will contain respectively the network
checkpoints and config, and the tensorboard run files. Each experiment
will be identified by and unique timestamped name.

TODO: when testing, reload the appropriate configuration