Fix a problem with parameter override in tests
Two kind of overrides happen before starting the testing procedure:
first, the train configuration is examined to make sure that its
parameters match the one from the test configuration; if it is not the
case, the train parameter override the test ones. Then the noise
parameters are overridden with default ones if not given. Before this
commit the procedure was inverted, so the wrong default noise params
might have been used.